Rename/move browser index.js into lib/browser.js
