+++ lib/browser.js
@@ -1,16 +1,15 @@
 'use strict';
+const HttpSoftAP = require('./HttpSoftAP');
+const TcpSoftAP = require('./TcpSoftAP');
+const SoftAP = require('./SoftAP.js');
 
-var HttpSoftAP = require('./lib/HttpSoftAP');
-var TcpSoftAP = require('./lib/TcpSoftAP');
-var SoftAP = require('./lib/SoftAP.js');
-
-var defaultPortMapping = {
+const defaultPortMapping = {
 	tcp: 5609,
 	http: 80
 };
 
 function SoftAPSetup(options) {
-	var opts = SoftAP.defaultOptions();
+	const opts = SoftAP.defaultOptions();
 	opts.protocol = 'tcp';
 
 	SoftAP.assign(opts, options);
